update DKVideoPlayer to https://github.com/Doikki/DKVideoPlayer commit 5203b55fcd01207b44e76e14f2730ced7cbdf5ee
